requirements.txt
hyperliquid-python-sdk>=0.9.0
eth-account>=0.11.0
requests>=2.31.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
